+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## WIP
+
+### Added
+
+- Support for handling ARM u-boot, firmware and overlays for Raspberry Pi (Workaround)
+
 ## [0.0.14]
 
 ### Fixed